pyproject.toml
[project]
name = "game-session-sync"
version = "0.1.0"
description = ""
authors = [{ name = "aloni636", email = "elyasaf.aloni@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<4" # notion-client requires <4 for some reason...
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "notion-client (>=2.4.0,<3.0.0)",
    "pydrive2 (>=1.21.3,<2.0.0)",
    "win11toast (>=0.35,<0.36)",
    "watchdog (>=6.0.0,<7.0.0)",
    "wmi (>=1.5.1,<2.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "pywin32 (>=311,<312)",
    "imagehash (>=4.3.2,<5.0.0)",
    "mss (>=10.1.0,<11.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "aiogoogle (>=5.17.0,<6.0.0)",
    "dacite (>=1.9.2,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = true

[tool.poetry.group.notebook.dependencies]
ipykernel = "^6.30.1"

[tool.black]

[tool.isort]
profile = "black"
known_pywin32 = [
    "ctypes",
    "psutil",
    "pythoncom",
    "win32gui",
    "win32process",
    "psutil",
    "wmi",
]
sections = [
    "FUTURE",
    "STDLIB",
    "PYWIN32",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
